Ran tests, removed unneeded files, renamed files, adjusted tests, view, and urls
