+++ cs_app/templates/one_page.html
@@ -47,7 +47,7 @@
                         </div>
                         <p>Change Database</p>
                     </div>
-                    <div class="page-link" onclick="redirectTo('/account_information_sub_page/')">
+                    <div class="page-link" onclick="redirectTo('/account_information/')">
                         <div class="symbol-sm flex-center">
                             <b>A</b>
                         </div>
