+++ cs_app/urls.py
@@ -27,13 +27,12 @@ urlpatterns = [
     path('get_db_info/', change_database_views.get_db_info_view, name='get_db_info'),
 
     # Change account page and functions
-    path('account_information/', account_information_views.account_information_view, name='change_account'),
+    path('account_information/', account_information_views.account_information_view, name='account_information'),
     path('account_information/update_name/', account_information_views.update_name_view, name='update_name'),
     path('account_information/update_email/', account_information_views.update_email_view, name='update_email'),
     path('account_information/update_phone/', account_information_views.update_phone_number_view, name='update_phone'),
     path('account_information/update_company/', account_information_views.update_company_view, name='update_company'),
     path('account_information/update_password/', account_information_views.update_password_view, name='update_password'),
-    path('account_information_sub_page/', account_information_views.account_information_sub_view, name='account_sub'),
 
 
 ]
